backend/requirements/requirements.txt
django==3.2.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
pillow==8.2.0